+++ NHibernate01/ORM_NHibernate/BusinessObjects/OnepTopologicallink.manualAdditions.cs
@@ -4,7 +4,7 @@ namespace ORM_NHibernate.BusinessObjects
     using System;
     using System.Collections.Generic;
 
-	public partial class OnepTopologicallink : BusinessBase<uint>
+	public partial class OnepTopologicallink : BusinessBase<long>
 	{
 		static OnepTopologicallink() { }
 		public virtual sbyte Discriminator { get; set; } = 1;
